requirements.txt
# ── UI ────────────────────────────────────────────────────────────────────────
streamlit==1.28.0

# ── LangChain ─────────────────────────────────────────────────────────────────
# langchain 0.1.x split integrations into separate packages
langchain>=0.1.0
langchain-core>=0.1.23
langchain-community>=0.0.10
langchain-groq>=0.1.3
langchain-text-splitters>=0.0.1
# ── Vector search ─────────────────────────────────────────────────────────────
faiss-cpu==1.7.4

# ── PDF extraction ────────────────────────────────────────────────────────────
pymupdf>=1.23.0

# ── Embeddings — text ─────────────────────────────────────────────────────────
sentence-transformers>=2.3.0

# ── Embeddings — image (CLIP) ─────────────────────────────────────────────────
transformers>=4.37.0               # CLIPModel + CLIPProcessor from HuggingFace
torch>=2.0.0                       # required by both transformers and sentence-transformers
pillow>=10.0.0                     # image loading / RGB conversion

# ── Numerics ──────────────────────────────────────────────────────────────────
numpy>=1.24.0

# ── LLM API ───────────────────────────────────────────────────────────────────
groq>=0.4.0

# ── Environment variables ─────────────────────────────────────────────────────
python-dotenv==1.0.0